pyproject.toml
[project]
name = "reduction-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "pillow",
    "scikit-learn",
    "numpy",
]
